✨ Update User Profile Form & Refactor Money Fields
- Addition of the possibility of updating the user profile without forcing the modification of the password.
- Addition of a confirmation field for the new password.
- Password security (only updates if a new value is provided).
- Improvement of the UI of the form with Tailwind CSS.
- Conversion of all fields in an interest in `Float` to avoid financial calculation errors.
- Use of `MoneyType` for` Wallet 'of `Usenty` and` Amount` in `Expensenty` and` incumentitith'.

✅ More flexible and ergonomic user profile.
✅ Better management of monetary values with `Moneytype`.

#Refactor #ux #Security #Database

+++ src/Entity/ExpenseEntity.php
@@ -27,7 +27,7 @@ class ExpenseEntity implements UserRelatedEntityInterface
 
     #[ORM\Column]
     #[Assert\Negative]
-    private ?int $amount = null;
+    private ?float $amount = null;
 
     #[ORM\Column(length: 255)]
     private ?string $name = null;
@@ -70,12 +70,12 @@ class ExpenseEntity implements UserRelatedEntityInterface
         return $this->id;
     }
 
-    public function getAmount(): ?int
+    public function getAmount(): ?float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): static
+    public function setAmount(float $amount): static
     {
         $this->amount = $amount;
